refactor: move KeyEvent.h to include and move KeyInfo to KeyEvent.h
