+++ RimeWithWeasel/RimeWithWeasel.cpp
@@ -269,15 +269,6 @@ DWORD RimeWithWeaselHandler::RemoveSession(WeaselSessionId ipc_id) {
   return 0;
 }
 
-namespace ibus {
-enum Keycode {
-  Escape = 0xFF1B,
-  XK_bracketleft = 0x005b, /* U+005B LEFT SQUARE BRACKET */
-  XK_c = 0x0063,           /* U+0063 LATIN SMALL LETTER C */
-  XK_C = 0x0043,           /* U+0043 LATIN CAPITAL LETTER C */
-};
-}
-
 void RimeWithWeaselHandler::UpdateColorTheme(BOOL darkMode) {
   RimeConfig config = {NULL};
   if (rime_api->config_open("weasel", &config)) {
@@ -323,7 +314,7 @@ BOOL RimeWithWeaselHandler::ProcessKeyEvent(KeyEvent keyEvent,
   Bool handled = rime_api->process_key(session_id, keyEvent.keycode,
                                        expand_ibus_modifier(keyEvent.mask));
   // vim_mode when keydown only
-  if (!handled && !(keyEvent.mask & (1 << 14))) {
+  if (!handled && !(keyEvent.mask & ibus::Modifier::RELEASE_MASK)) {
     bool isVimBackInCommandMode =
         (keyEvent.keycode == ibus::Keycode::Escape) ||
         ((keyEvent.mask & (1 << 2)) &&
