+++ WeaselIME/WeaselIME.vcxproj
@@ -482,7 +482,6 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="immdev.h" />
-    <ClInclude Include="KeyEvent.h" />
     <ClInclude Include="resource.h" />
     <ClInclude Include="stdafx.h" />
     <ClInclude Include="targetver.h" />
