+++ WeaselIME/WeaselIME.vcxproj.filters
@@ -44,9 +44,6 @@
     <ClInclude Include="immdev.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="KeyEvent.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
     <ClInclude Include="resource.h">
       <Filter>Header Files</Filter>
     </ClInclude>
